% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.data.continuous.example.R
\docType{data}
\name{cell.data.continuous.example}
\alias{cell.data.continuous.example}
\title{Example continuous cell data}
\format{A dataframe}
\usage{
cell.data.continuous.example
}
\description{
Example continuous cell data
}
\keyword{datasets}
