% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.data.discrete.example.R
\docType{data}
\name{cell.data.discrete.example}
\alias{cell.data.discrete.example}
\title{Example discrete cell data}
\format{A dataframe}
\usage{
cell.data.discrete.example
}
\description{
Example discrete cell data
}
\keyword{datasets}
