% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonotype.data.example.R
\docType{data}
\name{clonotype.data.example}
\alias{clonotype.data.example}
\title{Example clonotype data}
\format{
A dataframe
}
\usage{
clonotype.data.example
}
\description{
Example clonotype data
}
\keyword{datasets}
