% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTCellPack.R
\name{PlotTCellPack}
\alias{PlotTCellPack}
\title{Create T cell pack}
\usage{
PlotTCellPack(gliph = NULL, clonotype.data = NULL, cell.data = NULL,
  specificity.color = "#08306b", clonotype.color = "#4292c6",
  cell.color = "#9ecae1", color.gradient = "RdYlBu",
  line.color = "black", label.color = "black", label.size = 1,
  label = "none", legend = FALSE, cluster.size = 1)
}
\arguments{
\item{gliph}{Character vector providing the path to GLIPH_TCR_Table-convergence-groups.txt.}

\item{clonotype.data}{Optional 2-3 column data frame named "clonotype", "frequency" and/or "data".  If the second column
is "frequency", then the T cell clone size is drawn proportional to its frequency.  If the second column is "data", then the T cell clone
is colored according the variable (only discrete variables are supported).}

\item{cell.data}{Optional 3 column data frame named "clonotype", "cell", and "data".}

\item{specificity.color}{Character vector providing the color (single value) of the specificity circles.}

\item{clonotype.color}{Character vector providing the color (single value) of the clonotype circles.}

\item{cell.color}{Character vector providing the color (single value)  of the cell circles.}

\item{color.gradient}{Character vector providing the RColorBrewer color palette used color circles with discrete or continuous values.}

\item{line.color}{Character vector providing the color (single value) of the circle borders.}

\item{label.color}{Character vector providing the color (single value) of the labels.}

\item{label.size}{Integer corresponding to text size of the labels.}

\item{label}{Character vector indicating if labels should be displayed.  Options include "none", "specificity", "clonotype", "cell", and "data".}

\item{legend}{Logic indicating if legend should be displayed (TRUE) or not (FALSE).}

\item{cluster.size}{Minimum number of clonotypes per specificity group.}
}
\value{
Returns a T cell circle packing plot.
}
\description{
Creates a T cell pack from a gliph file, single-cell gene expression, or a combination of gliph and single-cell gene expression or gliph and T cell receptor sequencing.
}
\examples{
# T cell clone size proportional to frequency with GLIPH specificity groups
PlotTCellPack(gliph = gliph.example, clonotype.data = clonotype.data.example, legend = TRUE)

# T cell colored by continuous variable with GLIPH specificity groups
PlotTCellPack(gliph = gliph.example, cell.data = cell.data.continuous.example, legend = TRUE)

# T cell colored by discrete variable with GLIPH specificity groups
PlotTCellPack(gliph = gliph.example, cell.data = cell.data.discrete.example, legend = TRUE)

# T cell colored by discrete variable without GLIPH specificity groups
PlotTCellPack(cell.data = cell.data.discrete.example, legend = TRUE)

# T cell colored by continuous variable without GLIPH specificity groups
PlotTCellPack(cell.data = cell.data.continuous.example, legend = TRUE)

}
